template int/float pour les instances

+++ include/FileManager.hpp
@@ -11,4 +11,4 @@ using namespace std;
 
 void tokenize(std::string const &str, const char delim, vector<std::string> &out);
 int lecture_instance_alea( const string instanceName, vector< pair<float,float> > &res);
-int lecture_instance_tronc( const string instanceName, vector< pair<float,float> > &res);
+int lecture_instance_tronc( const string instanceName, vector< pair<int,int> > &res);
